feat: tidy example app, introduce address method
